Update README.md
I have added a brief description of the analyses but more needs to be added before we can move forward and I could use everyone's help with it. Also, I have some questions / notes. Those are in parentheses with (M:) next to them to indicate they're mine. Please let me know what you think.

I also modified our title slightly to be more specific.

I also added 2 more citations (from the proposal document that Daniel and I discussed). I changed the citations to all be numbered to follow the style we used beforebut we will need to update them because currently our citations are in 2 slightly different styles so we will need to fix them later.

Finally, I copied and pasted everything from our proposal that we needed in this new document so now it should all be in one place and we can continue to edit.